Move components into separate directory.
