+++ src/components/main-content/MainContent.scss
