+++ src/components/main-content/friends-list/FriendsList.scss
