+++ src/components/main-content/friends-list/index.ts
