+++ src/components/main-content/index.ts
